+++ src/Handlers/CreditcardPaymentHandler.php
@@ -30,6 +30,7 @@ class CreditcardPaymentHandler extends AsyncPaymentHandler
         string $version = null,
         array $gatewayInfo = []
     ): RedirectResponse {
+        $dataBag = $this->getRequestBag($dataBag);
         $paymentMethod = new Creditcard();
         return parent::pay(
             $transaction,