+++ src/Handlers/EpsPaymentHandler.php
@@ -29,6 +29,7 @@ class EpsPaymentHandler extends AsyncPaymentHandler
         string $version = null,
         array $gatewayInfo = []
     ): RedirectResponse {
+        $dataBag = $this->getRequestBag($dataBag);
         $paymentMethod = new Eps();
         return parent::pay(
             $transaction,