+++ src/Handlers/IdealProcessingPaymentHandler.php
@@ -29,6 +29,7 @@ class IdealProcessingPaymentHandler extends AsyncPaymentHandler
         string $version = null,
         array $gatewayInfo = []
     ): RedirectResponse {
+        $dataBag = $this->getRequestBag($dataBag);
         $paymentMethod = new IdealProcessing();
 
         if($issuer = $dataBag->get('bankMethodId')){