+++ src/Handlers/KbcPaymentHandler.php
@@ -29,6 +29,7 @@ class KbcPaymentHandler extends AsyncPaymentHandler
         string $version = null,
         array $gatewayInfo = []
     ): RedirectResponse {
+        $dataBag = $this->getRequestBag($dataBag);
         $paymentMethod = new Kbc();
         return parent::pay(
             $transaction,