+++ src/Handlers/WeChatPayPaymentHandler.php
@@ -29,6 +29,7 @@ class WeChatPayPaymentHandler extends AsyncPaymentHandler
         string $version = null,
         array $gatewayInfo = []
     ): RedirectResponse {
+        $dataBag = $this->getRequestBag($dataBag);
         $paymentMethod = new WeChatPay();
 
         $additional = [