+++ src/Helpers/CheckoutHelper.php
@@ -1075,17 +1075,34 @@ class CheckoutHelper
         return $this->orderRepository->search($orderCriteria, $context)->first();
     }
 
-    public function refundTransaction($order, $context, $item, $state, &$orderItems = '')
+    public function refundTransaction($order, $context, $item, $state, &$orderItems = '', $customRefundAmount = 0)
     {
         if (!$this->isBuckarooPaymentMethod($order)) {
             return;
         }
 
         $customFields = $this->getCustomFields($order, $context);
-
         $customFields['serviceName']            = $item['transaction_method'];
         $customFields['originalTransactionKey'] = $item['transactions'];
-        $amount                                 = $item['amount'];
+
+        $serviceName = (in_array($customFields['serviceName'], ['creditcard','creditcards', 'giftcards'])) ? $customFields['brqPaymentMethod'] : $customFields['serviceName'];
+
+        $amount = 0;
+        if ($customRefundAmount && !in_array($serviceName, ['afterpay', 'Billink', 'klarnakp' ])) {
+            $amount = $customRefundAmount;
+        } else {
+            if (!empty($orderItems) && is_array($orderItems)) {
+                foreach ($orderItems as $orderItem) {
+                    if (isset($orderItem['totalAmount'])) {
+                        $amount = $amount + $orderItem['totalAmount'];
+                    }
+                }
+            }
+        }
+
+        if ($amount <= 0) {
+            $amount = $item['amount']; //backward compatibility only or in case no $orderItems was passed
+        }
         $currency                               = !empty($item['currency']) ? $item['currency'] : 'EUR';
 
         if ($amount <= 0) {
@@ -1099,8 +1116,6 @@ class CheckoutHelper
         if (!empty($customFields['refunded']) && ($customFields['refunded'] == 1)) {
             return ['status' => false, 'message' => 'This order is already refunded'];
         }
-
-        $serviceName = (in_array($customFields['serviceName'], ['creditcard','creditcards', 'giftcards'])) ? $customFields['brqPaymentMethod'] : $customFields['serviceName'];
 
         $request = new TransactionRequest;
         $request->setServiceAction('Refund');
@@ -2110,4 +2125,17 @@ class CheckoutHelper
             return $statusCodeAddErrorMessage[$statusCode] ?? '';
     }
 
+    public function getSession()
+    {
+        return $this->session;
+    }
+
+    public function areEqualAmounts($amount1, $amount2)
+    {
+        if ($amount2 == 0) {
+            return $amount1 == $amount2;
+        } else {
+            return abs((floatval($amount1) - floatval($amount2)) / floatval($amount2)) < 0.00001;
+        }
+    }
 }