+++ src/Resources/app/storefront/src/main.js
@@ -1,6 +1,8 @@
 import BuckarooPaymentCreditcards from './creditcards/buckaroo-payment.creditcards';
 import BuckarooPaymentHelper from './helper/buckaroo-payment.helper';
+import BuckarooPaymentValidateSubmit from './helper/buckaroo-validate';
 
 const PluginManager = window.PluginManager;
+PluginManager.register('BuckarooPaymentValidateSubmit', BuckarooPaymentValidateSubmit);
 PluginManager.register('BuckarooPaymentCreditcards', BuckarooPaymentCreditcards);
 PluginManager.register('BuckarooPaymentHelper', BuckarooPaymentHelper);